refactor: Timer에서 타이머 작동 부분 코드 분리
